initial connection between telegraf, influx, grafana works
We have initially tested the connection between telegraf, grafana, influxdb by pulling cpu data. We also have the mqtt part setup where data is being read from the csv file and published to mqtt. From here onwards, we need to connect the mqtt and telegraf and replace the cpu data with the mqtt data in influx and show it in grafana.

Key Points:
->Please replace the telegraf.config file with appropriate config file for connecting with mqtt
->Uncomment the mqtt part in the .env part.